Changes days param to a double so sub 1 day ranges can be requested

+++ src/commonMain/kotlin/drewcarlson/coingecko/CoinGeckoClient.kt
@@ -99,7 +99,7 @@ interface CoinGeckoClient {
     suspend fun getCoinMarketChartById(
         id: String,
         vsCurrency: String,
-        days: Int
+        days: Double
     ): MarketChart
 
     @Throws(CoinGeckoApiException::class, CancellationException::class)
